styling: :hammer: move symlink arrow to styling mod and use styling
fix https://github.com/Peltoche/lsd/issues/409
